chore: 응집도를 위해 constants 상수들을 각 역할에 맞는 디렉토리에 이동
