+++ src/domain/getRandomNumber.js
@@ -1,4 +1,4 @@
-import { LOTTO } from '../constants/messages.js';
+import { LOTTO } from './lottoConstants.js';
 
 export const getRandomNumber = () => {
   return Math.floor(Math.random() * LOTTO.MAX_RANDOM_NUMBER) + LOTTO.MIN_RANDOM_NUMBER;
