+++ src/domain/lottoConstants.js
@@ -1,0 +1,17 @@
+export const LOTTO = Object.freeze({
+  MIN_PURCHASE_PRICE: 1000,
+  MAX_RANDOM_NUMBER: 45,
+  MIN_RANDOM_NUMBER: 1,
+  MAX_LENGTH: 6,
+  FIVE_WITH_BONUS_MATCH_IDX: 7,
+  SIX_MATCH: 6,
+  FIVE_WITH_BONUS_MATCH: 5.5,
+  FIVE_MATCH: 5,
+  FOUR_MATCH: 4,
+  THREE_MATCH: 3,
+  PRIZE_OF_SIX_MATCH: 2_000_000_000,
+  PRIZE_OF_FIVE_WITH_BONUS_MATCH: 30_000_000,
+  PRIZE_OF_FIVE_MATCH: 1_500_000,
+  PRIZE_OF_FOUR_MATCH: 50_000,
+  PRIZE_OF_THREE_MATCH: 5_000,
+});
